Move exportItemReport out of main
